Initial commit
In attempting to evaluate whether or not tests are actually failing
because of the test itself and not because previously run test did bad
things, I decided to write some scripts that take each test from a
generated list of tests defined by a "test group" and run each, one at a
time.

If a test fails, it is recorded in a list of failing tests and the cleanup
scripts are run, redirecting stdout to the test result file. The cleanups
are also run before running each test just in case there is something
laying around that would mess it up.